moved isloggedIn from utils to middleware
